+++ core/src/main/kotlin/in/specmatic/core/wsdl/payload/ComplexTypedSOAPPayload.kt
@@ -1,8 +1,8 @@
 package `in`.specmatic.core.wsdl.payload
 
+import `in`.specmatic.core.pattern.TYPE_ATTRIBUTE_NAME
 import `in`.specmatic.core.value.toXMLNode
 import `in`.specmatic.core.wsdl.parser.SOAPMessageType
-import `in`.specmatic.core.wsdl.parser.buildXmlDataForComplexElement
 import `in`.specmatic.core.wsdl.parser.message.AttributeElement
 
 data class ComplexTypedSOAPPayload(
@@ -18,4 +18,16 @@ data class ComplexTypedSOAPPayload(
         val body = soapMessage(toXMLNode(xml), namespaces)
         return listOf("And ${soapMessageType.qontractBodyType}-body\n\"\"\"\n${body.toPrettyStringValue()}\n\"\"\"")
     }
+}
+
+fun buildXmlDataForComplexElement(
+    nodeName: String,
+    qontractTypeName: String,
+    attributes: List<AttributeElement>
+): String {
+    val attributeString = attributes.joinToString(" ") {
+        "${it.nameWithOptionality}=\"${it.type}\""
+    }
+
+    return "<${nodeName} $TYPE_ATTRIBUTE_NAME=\"$qontractTypeName\" $attributeString />"
 }
